refactor(kafka-message-generator): share ApiData base between generators
ApiKeysGenerator and ApiMessageTypeGenerator each declared a near-identical
ApiData nested class to pair request/response specs per API key. Pulled
the common apiKey/requestSpec/responseSpec/name() logic into ApiSpec so
each generator only owns what it adds on top.
